tests: Move fw_cfg to integration_fw_cfg.rs
Extract mod fw_cfg from integration.rs into a standalone
test binary integration_fw_cfg.rs. This follows the
pattern established by integration_cvm.rs.
